# Add TestMetadataCache to improve CompositeWorkItem build speed