@lahma 2023-02-27
> Thanks for providing this @lahma. I'm a bit tired today, so I cannot determine if we will run into some of the same problems as we did for the other cache - see e.g. #4104 for the solution.

Thanks for the review. At least in this case the usage in in single point so if the attribute fails at first use it should throw at that point. Consecutive calls of course get the cached value but like discussed in one of the threads that kind of stateful attribute usage would be considered invalid.