@OsirisTerje 2023-04-06
@mikkelbu Have you had time to look into this?   Would be nice to have it merged in.