@lahma 2023-04-13
@manfred-brands I've done the fixes you requested, I squashed to single commit keep cleaner, PR title and commit text now reflect the new name `TestMetadataCache`.